Split the ReflectionHelpers out in to a partial class implementation
